Better error messaging on install if cached releases.json is stale
On `hashi-env install`, there previously was no error message saying why
it couldn't find the designated version. First issue with the `grep -Fe`
line where it filtered down to the given version fragment. That's now
handled upstream by the shell function we're calling.

Next was adding a bit of retry logic and fetching an updated
`releases.json` if it isn't able to find the specified version. If
_neither_ is successful, _then_ we raise a human-facing error full with
an actual message. Novel idea, no?